Move duplicates to MachineBlock and MachineBlockEntity
New MachineInterface
