+++ src/main/java/com/skilles/cannacraft/blocks/machines/strainAnalyzer/StrainAnalyzer.java
@@ -1,7 +1,6 @@
 package com.skilles.cannacraft.blocks.machines.strainAnalyzer;
 
 import com.skilles.cannacraft.blocks.machines.MachineBlock;
-import com.skilles.cannacraft.blocks.machines.seedCrosser.SeedCrosserEntity;
 import com.skilles.cannacraft.registry.ModEntities;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
@@ -9,53 +8,24 @@ import net.minecraft.block.entity.BlockEntityTicker;
 import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.item.ItemStack;
-import net.minecraft.particle.DustParticleEffect;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
-import net.minecraft.util.ItemScatterer;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Direction;
-import net.minecraft.util.math.Vec3f;
 import net.minecraft.world.BlockView;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
-import java.util.Random;
 
 public class StrainAnalyzer extends MachineBlock {
 
     public StrainAnalyzer(Settings settings) {
         super(settings);
-        setDefaultState(getStateManager().getDefaultState().with(FACING, Direction.NORTH).with(ACTIVE, false));
     }
-
-    @Override
-    public void randomDisplayTick(BlockState state, World world, BlockPos pos, Random random) {
-        if (state.get(ACTIVE)) {
-
-            Direction direction = state.get(FACING);
-            Direction.Axis axis = direction.getAxis();
-            double h = random.nextDouble() * 0.6D - 0.3D;
-            double i = axis == Direction.Axis.X ? direction.getOffsetX() * 0.52D : h;
-            double j = random.nextDouble() * 6.0D / 16.0D;
-            double k = axis == Direction.Axis.Z ? direction.getOffsetZ() * 0.52D : h;
-
-            float r = random.nextFloat();
-            float g = random.nextFloat();
-            float b = random.nextFloat();
-
-            world.addParticle(new DustParticleEffect(new Vec3f(r, g, b), .7f), pos.getX() + 0.5 + i,
-                    pos.getY() + j, pos.getZ() + 0.5 + k, 0, 0, 0);
-        }
-    }
-
 
     @Nullable
     @Override
-    public BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
-        return new StrainAnalyzerEntity(pos, state);
-    }
+    public BlockEntity createBlockEntity(BlockPos pos, BlockState state) { return new StrainAnalyzerEntity(pos, state); }
 
     @Nullable
     @Override
@@ -66,17 +36,5 @@ public class StrainAnalyzer extends MachineBlock {
     @Override
     public void appendTooltip(ItemStack stack, @Nullable BlockView world, List<Text> tooltip, TooltipContext options) {
         tooltip.add(Text.of("Analyzes unidentified seeds").shallowCopy().formatted(Formatting.GOLD));
-    }
-    @Override
-    public void onStateReplaced(BlockState state, World world, BlockPos pos, BlockState newState, boolean moved) {
-        if (state.getBlock() != newState.getBlock()) {
-            BlockEntity blockEntity = world.getBlockEntity(pos);
-            if (blockEntity instanceof SeedCrosserEntity) {
-                ItemScatterer.spawn(world, pos, (SeedCrosserEntity)blockEntity);
-                // update comparators
-                world.updateComparators(pos,this);
-            }
-            super.onStateReplaced(state, world, pos, newState, moved);
-        }
     }
 }
