+++ src/main/java/com/skilles/cannacraft/blocks/machines/weedExtractor/WeedExtractorEntity.java
@@ -1,5 +1,6 @@
 package com.skilles.cannacraft.blocks.machines.weedExtractor;
 
+import com.skilles.cannacraft.blocks.machines.MachineBlock;
 import com.skilles.cannacraft.blocks.machines.MachineBlockEntity;
 import com.skilles.cannacraft.registry.ModEntities;
 import com.skilles.cannacraft.registry.ModItems;
@@ -9,14 +10,9 @@ import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
-import net.minecraft.inventory.Inventories;
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NbtCompound;
 import net.minecraft.screen.PropertyDelegate;
 import net.minecraft.screen.ScreenHandler;
-import net.minecraft.sound.SoundCategory;
-import net.minecraft.sound.SoundEvent;
-import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
@@ -62,18 +58,16 @@ public class WeedExtractorEntity extends MachineBlockEntity {
             }
         };
     }
-
-
     public static void tick(World world, BlockPos pos, BlockState state, WeedExtractorEntity blockEntity) {
         if (world == null || world.isClient) return;
-        if(isNextTo(world, pos, Blocks.GLOWSTONE) && blockEntity.powerStored < blockEntity.getMaxStoredPower()) {
+        if (isNextTo(world, pos, Blocks.GLOWSTONE) && blockEntity.powerStored < blockEntity.getMaxStoredPower()) {
             blockEntity.addEnergy(2);
             markDirty(world, pos, state);
         }
         if (blockEntity.isWorking()) {
             if (!world.isReceivingRedstonePower(pos)) {
                 processTick(blockEntity); // playSound is called here
-                state = state.with(WeedExtractor.ACTIVE, true);
+                state = state.with(MachineBlock.ACTIVE, true);
                 world.setBlockState(pos, state, Block.NOTIFY_ALL);
                 markDirty(world, pos, state);
             }
@@ -86,48 +80,13 @@ public class WeedExtractorEntity extends MachineBlockEntity {
                 markDirty(world, pos, state);
             }
         } else if (canCraft(blockEntity.inventory) && blockEntity.powerStored != 0) { // start if has power
-                blockEntity.processingTime = 1;
+            blockEntity.processingTime = 1;
         } else { // when no items or can't craft
-                blockEntity.processingTime = 0;
-                state = state.with(WeedExtractor.ACTIVE, false);
-                world.setBlockState(pos, state, Block.NOTIFY_ALL);
-            }
-            markDirty(world, pos, state);
+            blockEntity.processingTime = 0;
+            state = state.with(MachineBlock.ACTIVE, false);
+            world.setBlockState(pos, state, Block.NOTIFY_ALL);
         }
-    private void playSound() {
-        World world  = getWorld();
-        SoundEvent runSound = SoundEvents.BLOCK_FIRE_AMBIENT;
-        assert world != null;
-        if(!world.isClient) {
-            if(this.processingTime % 25 == 0) {
-                if (this.processingTime != timeToProcess) {
-                    world.playSound(
-                            null, // Player - if non-null, will play sound for every nearby player *except* the specified player
-                            pos, // The position of where the sound will come from
-                            runSound, // The sound that will play, in this case, the sound the anvil plays when it lands.
-                            SoundCategory.BLOCKS, // This determines which of the volume sliders affect this sound
-                            0.15f, //Volume multiplier, 1 is normal, 0.5 is half volume, etc
-                            0.5f // Pitch multiplier, 1 is normal, 0.5 is half pitch, etc
-                    );
-                } else {
-                    world.playSound(
-                            null, // Player - if non-null, will play sound for every nearby player *except* the specified player
-                            pos, // The position of where the sound will come from
-                            runSound, // The sound that will play, in this case, the sound the anvil plays when it lands.
-                            SoundCategory.BLOCKS, // This determines which of the volume sliders affect this sound
-                            0.15f, //Volume multiplier, 1 is normal, 0.5 is half volume, etc
-                            2f // Pitch multiplier, 1 is normal, 0.5 is half pitch, etc
-                    );
-                }
-            }
-        }
-    }
-    public boolean isWorking() {
-        if(needsPower) {
-            return processingTime != 0 && powerStored != 0;
-        } else {
-            return processingTime != 0;
-        }
+        markDirty(world, pos, state);
     }
     public static boolean canCraft(DefaultedList<ItemStack> inventory) {
             ItemStack stack = inventory.get(1);
@@ -139,22 +98,18 @@ public class WeedExtractorEntity extends MachineBlockEntity {
             }
         return false;
     }
-    private static void processTick(WeedExtractorEntity blockEntity) {
-        blockEntity.processingTime++;
-        if(blockEntity.needsPower) blockEntity.useEnergy(1 * blockEntity.powerMultiplier);
-        blockEntity.playSound();
-    }
-    public static void craft(DefaultedList<ItemStack> inventory) {
-            ItemStack stack = inventory.get(1);
-            ItemStack outputSlot = inventory.get(0);
-            ItemStack output = StrainUtil.getOutputStack(stack);
-            if(outputSlot.isEmpty()) {
-                inventory.set(0, output);
-            }
-            else if (outputSlot.isOf(output.getItem())) {
-                outputSlot.increment(1);
-            }
-            stack.decrement(1);
+    protected static int craft(DefaultedList<ItemStack> inventory) {
+        ItemStack stack = inventory.get(1);
+        ItemStack outputSlot = inventory.get(0);
+        ItemStack output = StrainUtil.getOutputStack(stack);
+        if(outputSlot.isEmpty()) {
+            inventory.set(0, output);
+        }
+        else if (outputSlot.isOf(output.getItem())) {
+            outputSlot.increment(1);
+        }
+        stack.decrement(1);
+        return 0;
     }
 
     @Nullable
@@ -162,26 +117,8 @@ public class WeedExtractorEntity extends MachineBlockEntity {
     public ScreenHandler createMenu(int syncId, PlayerInventory inv, PlayerEntity player) {
         return new WeedExtractorScreenHandler(syncId, inv, this, this.propertyDelegate);
     }
-
-    @Override
-    public NbtCompound writeNbt(NbtCompound nbt) {
-        super.writeNbt(nbt);
-        nbt.putInt("processingTime", this.processingTime);
-        nbt.putInt("powerStored", this.powerStored);
-        Inventories.writeNbt(nbt, this.inventory);
-        return nbt;
-    }
-
-    @Override
-    public void readNbt(NbtCompound nbt) {
-        super.readNbt(nbt);
-        this.inventory = DefaultedList.ofSize(this.inventory.size(), ItemStack.EMPTY);
-        Inventories.readNbt(nbt, this.inventory);
-        this.processingTime = nbt.getInt("processingTime");
-        this.powerStored = nbt.getInt("powerStored");
-    }
     @Override
     public boolean canInsert(int slot, ItemStack stack, @Nullable Direction dir) {
-        return stack.isOf(ModItems.WEED_SEED);
+        return stack.isOf(ModItems.WEED_FRUIT);
     }
 }
